+++ r/NEWS.md
@@ -23,6 +23,7 @@
 * `write_feather`, `write_arrow` and `write_parquet` now return their input
 similar to `write_*` functions from `readr` (#6387, @boshek)
 * Dataset filtering is now correctly supported for all Arrow date/time/timestamp column types.
+* Can now infer the type of an R `list` and create a ListArray when all list elements are the same type (#6275, @michaelchirico)
 
 # arrow 0.16.0